content/canvas/test/test_2d.imageData.create.large.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.create.large - bug 433004</title>
<!-- Testing: createImageData() works for sizes much larger than the canvas -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;

var imgdata = ctx.createImageData(1000, 2000);
ok(imgdata.data.length == imgdata.width*imgdata.height*4, "imgdata.data.length == imgdata.width*imgdata.height*4");
ok(imgdata.width < imgdata.height, "imgdata.width < imgdata.height");
ok(imgdata.width > 0, "imgdata.width > 0");
var isTransparentBlack = true;
for (var i = 0; i < imgdata.data.length; i += 7813) // check ~1024 points (assuming normal scaling)
    if (imgdata.data[i] !== 0)
        isTransparentBlack = false;
ok(isTransparentBlack, "isTransparentBlack");

SimpleTest.finish();

});
</script>
